bin/MGPGtools/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mgpgtools"
dynamic = ["version"]
readme = "README.md"
#license = "MIT"
requires-python = ">=3.4"
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "biopython",
    "gfapy",
    "pandas",
    "prettytable",
    "toytree",
]

[project.scripts]
panTools = "main:main"

[tool.hatch.version]
path = "info.py"

[tool.hatch.build.targets.sdist]
include = [
    "/utils",
    "*.py",
]

[tool.hatch.build.targets.wheel]
include = [
    "/utils",
    "*.py",
]
